requirements.txt
# ========================================
# Swing Quant Workflow - 依赖清单
# ========================================

# === 核心依赖 ===
loguru>=0.7.0              # 日志系统
pyyaml>=6.0                # YAML 配置解析

# === LLM 模型客户端 ===
# 至少安装其中一个：
anthropic>=0.39.0          # Anthropic Claude API
openai>=1.54.0             # OpenAI / DMXapi (OpenAI兼容)

# === 数据处理 ===
pandas>=2.0.0              # 数据分析
numpy>=1.24.0              # 数值计算

# === HTTP 请求 ===
requests>=2.31.0           # API 请求
httpx>=0.27.0              # 异步 HTTP 客户端

# === 图片处理（可选，Agent3需要）===
pillow>=10.0.0             # 图片预处理

# === JSON Schema 验证 ===
jsonschema>=4.17.0         # JSON Schema 验证

# === 类型提示 ===
typing-extensions>=4.5.0   # 类型注解扩展

# === 开发工具（可选）===
# pytest>=7.4.0            # 单元测试
# black>=23.0.0            # 代码格式化
# flake8>=6.0.0            # 代码检查